requirements.txt
transformers
torch
datasets
pandas
openpyxl
python-docx
accelerate
scikit-learn
